Rework mocks in tests
