+++ jest.config.js
@@ -8,7 +8,11 @@ const createJestConfig = nextJest({
 
 // Add any custom config to be passed to Jest
 const customJestConfig = {
-  setupFilesAfterEnv: ["<rootDir>/jest.setup.js", "<rootDir>/tools/setEnvVars.js"],
+  setupFilesAfterEnv: [
+    "<rootDir>/jest.setup.js",
+    "<rootDir>/jest.mock.js",
+    "<rootDir>/tools/setEnvVars.js",
+  ],
   testEnvironment: "jest-environment-jsdom",
   collectCoverageFrom: ["src/**/*.{ts,tsx}"],
 };
